refactor: rename RoomVote to Vote
