--- piikii-application/src/main/kotlin/com/piikii/application/domain/roomvote/RoomVote.kt
@@ -1,9 +1,0 @@
-package com.piikii.application.domain.roomvote
-
-import java.util.UUID
-
-class RoomVote(
-    val userId: UUID,
-    val roomPlaceId: Long,
-    val content: String,
-)
