+++ piikii-application/src/main/kotlin/com/piikii/application/domain/vote/Vote.kt
@@ -1,0 +1,9 @@
+package com.piikii.application.domain.vote
+
+import java.util.UUID
+
+class Vote(
+    val userId: UUID,
+    val roomPlaceId: Long,
+    val result: VoteResult,
+)
