+++ piikii-output-persistence/postgresql/src/main/kotlin/com/piikii/output/persistence/postgresql/persistence/repository/PlaceVoteRepository.kt
@@ -1,0 +1,6 @@
+package com.piikii.output.persistence.postgresql.persistence.repository
+
+import com.piikii.output.persistence.postgresql.persistence.entity.VoteEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface PlaceVoteRepository : JpaRepository<VoteEntity, Long>
