--- piikii-output-persistence/postgresql/src/main/kotlin/com/piikii/output/persistence/postgresql/persistence/repository/RoomVoteRepository.kt
@@ -1,6 +1,0 @@
-package com.piikii.output.persistence.postgresql.persistence.repository
-
-import com.piikii.output.persistence.postgresql.persistence.entity.RoomVoteEntity
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface RoomVoteRepository : JpaRepository<RoomVoteEntity, Long>
